Re-implemented direnv-diff and direnv-dump in Go
For two reasons:

1) direnv is still not fast enough for me. I see slow starting consoles
and I suspect it is due to ruby loading a tons of files.

2) I want to remove the ruby dependency. Ideally I would use bash but
it's not powerful enough. Or C but it would take me ages to build
something reliable. Go can cross-compile and is statically linked.

Timing the whole test suite does not make any difference with the Go
implementation.