Correct potential false positive load
It is possible to falsely report a successful load of NodeJS when
the specified version was not actually loaded but you've got another
installed version of node (i.e. brew install node). This fix corrects
that false positive by checking whether there is a `node` binary
(executable) within the `$node_prefix`.