Build-time bash path
This change allows to hard-code the bash path into direnv at compile
time. This avoiding a lookup on the $PATH on every evaluation of direnv.

To do this run `make BASH_PATH=$(readlink -f $(which bash))`.

There is no measurable difference on a SSD.

This is going to be useful in nix packaging to provide bash path as an
immutable build dependency.