stdlib: configurable direnv_layout_dir
Introduce the `$direnv_layout_dir` variable to let the user select
out-of-folder layouts. The main purpose of making this configurable is
to allow the user to store the `.direnv` folder outside of the
project's directory. For example if it interracts badly with backups,
source control or the IDE.

The variable can be changed in the `.envrc` file or in the global
`~/.config/direnv/direnvrc` file. In that case on might want to hash the
project path like that:

```
: ${XDG_CACHE_HOME:=$HOME/.cache}
pwd_hash=$(echo -n $PWD | shasum | cut -d ' ' -f 1)
direnv_layout_dir=$XDG_CACHE_HOME/direnv/layouts/$pwd_hash
```