stdlib: source the direnvrc directly
This changes direnv's behaviour.

It will no longer track the direnvrc for changes or print when it's
loaded. $PWD is also now the .envrc's folder in the direnvrc context.

It turns out to be useful to know the .envrc's current directory inside
of the direnvrc context. Also logging that the direnvrc has been loaded
is not super interesting.