Overhaul the build system
== Build with Nix

Add all the build dependencies to CI with Nix. Nix is
not a requirement to build direnv but you need to do specify the
dependencies yourself otherwise.

== Allow to checkout direnv everywhere

Handle our own GOPATH in the Makefile. direnv is still buildable with
`go get github.com/direnv/direnv` but should also work if go is just
installed in the system and the user checks out direnv anywhere and runs
`make`

== Better Travis-CI testing

Test for more shells by default and also for formatting by default.

Avoid unnecessary regressions. Some shells like elvish and tcsh are
broken and will need to be fixed in the future.

Test on Darwin

== Format all the things

stdlib.sh is formatted with shfmt

== Misc

Makefile: use lower-case variables. When a variable is not exported and
not meant to be changed by the user, use lower-case variables

Rename Makefile to GNUmakefile as it wasn't tested with the other make
implementations.

Ensure that bash is being used with make

Replace usage of `which` with shell builtins when possible.