Add shell_specific functionality
The `shell_specific()` function is a higher-level version of
`quote()`: it let's one send to the shell not only a command
to do "on loading" the env, but also a command that can compute
the diff with respect to the current env and record it, in the
`DIRENV_ON_UNLOAD_XXX` variable, to be executed "on unloading".

From the point of view of `direnv export xxx`, the `DIRENV_ON_UNLOAD_XXX`
variables are similar to `DIRENV_DIFF`. The protocol that is
implemented is:

  - When running `direnv export xxx`, if we are unloading, only
    `DIRENV_ON_UNLOAD_xxx` is considered.

  - `DIRENV_ON_UNLOAD_xxx` is a comma-separated list of
    shell-specific commands, encoded using gzenv.

  - When the decoded commands are output, they are separated
    by newlines