Allow for aliases to be exported
We introduce the `use aliases` functionality in
the stdlib. Once activated, the `alias` command
will effectively set aliases on bash, zsh, tcsh
and fish by way of shell-specific code. It is
implemented by aliasing "alias" to a function
that mimics the behaviour of "alias" under
"bash" but also emits the corresponding
shell-specific code.

On each shell, a command is run to find out the
current definition for the alias (in case it is
being redefined)  and an "on-unload" action is
registered that will restore the original
definition or remove the alias in case it is
new.

In the case of tcsh, this was tricky to test
since in order to decide if, on unload, we
need to execute a `unalias foo` or an `alias
foo old-value`, we need to run a command that
in some cases will finish with a non-zero
value to be able to perform the "if". However,
`tcsh -e` will fail immediately when this
happens (the alternative would be to use
backticks, but they can't be easily nested),
etc. So we end up adding another test script
for tcsh to be run with tcsh, for this type
of scenarios...