test-{bash/zsh}: Make direnv_eval match hook code
In the hook code, we do essentially `eval "$(direnv export bash)"`,
while on the test code we were doing `eval $(direnv export bash)`.
These are not equivalent, since spaces in the output of `direnv export
bash` will break the output; so that everything that comes after gets
interpreted outside of the `eval`.