Use source instead of eval on fish hook
At the moment, fish has no equivalent of the `"$(cmd)"` substitution
of bash, zsh, etc. In particular, `(cmd)` will split the output of
`cmd` on newlines. That means that `eval (direnv export fish)`,  will
first run `direnv export fish`, then split it on newlines and finally
call `eval` with one argument per line. The behaviour of `eval` is to
concatenate all its arguments with *spaces* (not newlines) and then
evaluate that.

This is a bug waiting to happen (e.g., if the value of an exported
value contains a `\n` it could be silently replaced by a space), and
would also prevent us from implementing `shell_specific` for `fish`.

In fish, though, `eval` is just a wrapper around `source`, which can
actually take it's input from stdin. So we can simply pipe the output
of `direnv export fish` to `source`.