Simplify direnv_load and make it work even when the command crashes.
Previously a call to, for example, `direnv_load true` would cause direnv to
hang. This is because the command didn't include a call to `direnv dump`, hence
nothing ever wrote to the far end of the named pipe that had been created to
receive the output from the given command, `true`.

Another example:

  direnv_load bash -c 'do_something_that_breaks && direnv dump'

In this instance we've rememebered to call `direnv dump`, but it is never called
because `do_something_that_breaks` returns non-zero.