config: add [global] section
Because there is not way to close a section in toml, it's better to
enclose each value in a section. Otherwise top-level values can only be
put in the top of the file and it's confusing.

Here we add a new [global] section to put all the top-level values.
Backward-compatibility is kept with the old values as well.