test: elvish: Fix evaluation function
The direnv hook for elvish was changed in commit ab4d188d ("shell: fix
elvish hook"). It seems that there is a change in shell behaviour that
is being handled by the commit. (It's hard to understand the reason
for the change as there are no details in the commit log).

Recently, build failures attributable to failing elvish tests were
reported in debian (See Bug #952168). On further investigation, it
seems a similar change to the elvish test script is needed.