Draft Windows support (fix for #594 and related)
* Fix for #594 (write error to fd 3 on Windows)
* Fix wrong SelfPath for 'direnv hook' on Windows
* Fix `$PATH` corruption across process boundaries

The PATH corruption issue is fixed by adding an extra command line
argument, `UNIX_PATH`, to `direnv export`, `direnv dump`, and
stdlib's `__main__`:

If it is non-empty and *different* from the value of `$PATH`, it is
used in place of `$PATH` for purposes of generating the exported
environment.  Shells supply this argument from their hook function
to bypass the conversion between Windows and Unix-style paths that
takes place when changing between shells and Go programs on Windows,
thereby preserving symlinks, case, etc. that would otherwise be lost
in translation.

Internally, a DIRENV_PLATFORM_PATH key is added to the logical
environment processed by `direnv export`, `direnv dump`, and
`direnv exec`, and its changes are tracked in DIRENV_DIFF, but
it is never actually exposed as an environment variable to parent
or child processes, nor is it ever read from the OS environment.

(It's tracked in DIRENV_DIFF so that `direnv exec` can know the
previous OS path, though at the moment that's not very useful
because syscall.Exec doesn't work on Windows and I don't know
what a proper replacement would be.  So for the moment, `direnv
exec` still doesn't work on Windows.)