Add on_git_branch command to detect whether a specific git branch is checked out (#702)
When working on GitHub projects, I've found it helpful to have env variables set-up when I'm inside a specific branch. For example, sometimes I want to change the merge-base my editor uses to display signifiers. I can accomplish this with the new command like so:

    if on_git_branch child_changes; then
      export MERGE_BASE_BRANCH=parent_changes
    fi

To detect movement between branches, the .git/HEAD file is watched with watch_file. This is only done when a branch name is specified, otherwise, no calls to watch_file are made.