GNUmakefile: change packaging. DESTDIR -> PREFIX, fish hook
If you are maintaining a direnv package, your packaging instructions
will have to change.

Rename DESTDIR to PREFIX. Semantically, PREFIX is used to declare where
the files will live on the file-system after installation and DESTDIR is
a temporary location used during packaging. This commit fixes that
semantic.

Update the fish hook and distribute it with the package. These can be
removed from the packaging instructions now.