Fix using PWD in .env files
This change introduces PWD support to the `dotenv` functionality, so
that it can resolve PWD in `.env` files.

When .env is loaded, PWD will be set to the same folder that .env file
can be found in variable can be used.
A user would also expect that the PWD variable is matching the folder
that the .env file resides in, and symlinks should not interfere with
that either.

Fixes #1051