stdlib: use_flake: don't keep old generations around
This is a pretty significant change for people who care about this, but
I'd argue that this is very rarely the case.

The two reasons I can see why one would like to keep an old generation
are
1. Enable rollbacks
2. Prevent garbage collection

Given that use_flake is always used with a flake.nix file and most
probably in a version controlled project, reverting flake.nix seems like
an easier way to roll back than fiddling with generations (which would
be overridden by direnv pretty soon anyways)

Garbage collection is a better reason imo, but still, actually going
back to old versions is probably relatively rare, and stuff isn't
collected before nix-collect-garbage is ran anyways.