Fix incorrect escape sequences during Loads under git-bash (#1085)
https://github.com/direnv/direnv/pull/975 had the unintended consequence
of causing the `RC::Load` call to sometimes fail when running under
git-bash (and some others Windows specific configs).

The `RC::Load` -> `cmd.Output()` calls end up interpreting
characters like \b or \t as escape sequences, which is problematic if
for paths like 'myProject\buckets\test'.

This patch addresses this by forcing the use of `/` as a separator, even
on Windows. This works fine in my testing in git-bash. This shouldn't
impact systems where `/` is already in use.

This also adds a test case that covers the behaviour being adjusted. In
order to run the bash tests on Windows, I had to disabled a number that
appear to not play nice -- but after hacking those out I was able to
verify the new test case.

Fixes:
- https://github.com/direnv/direnv/issues/1020
- https://github.com/direnv/direnv/issues/1079