stdlib: add use_flakepath function
Function is very similiar to use_flake, except it parses the PATH
variable of the nix environment and appends it.

Users of other shells that don't want to use "shellHook" to exec back
into their prefered shells, and only need PATH modified might appreciate
this.

In contradiction to use_flake we don't create a profile, if you use
nix-collect-garbage while inside a shell like this, you'll lose your
PATH entries.