Powershell: Delimit escaped envrionment variable names when setting
Without this many exports, such as `$env:'depsBuildTargetPropagated'='';`,
are not valid Powershell and result in an error being thrown:
```
Invoke-Expression -Command $export;
     |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     | Variable reference is not valid. ':' was not followed by a valid
     | variable name character. Consider using ${} to delimit the name
```

The quotes around escaped variable names cause Powershell to be confused
because it can't figure out what is part of the variable reference. So just
delimit the name to make it clear what's part of the name and what's part
of the value.